+++ src/components/player.js
@@ -1,0 +1,7 @@
+const createPlayer = (nameArg, signArg) => {
+  const name = nameArg;
+  const sign = signArg;
+  return { name, sign };
+};
+
+export default createPlayer;
