--- testing/index.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Tika Taka Toka</title>
-    <link rel="stylesheet" href="assets/style.css">
-    <link href="https://fonts.googleapis.com/css2?family=Rock+Salt&display=swap" rel="stylesheet">
-  </head>
-  <body>
-    <div class="main-container">
-      <header>
-        <h2 id="game-message">Tic Tac Toe</h2>
-        <button id="reset">Restart</button>
-      </header>
-      <div class="game-container">
-        <div class="buttons">
-          <input id="plr1" type="text" name="plr1" placeholder="Player 1 is O" />
-          <input id="plr2" type="text" name="plr2" placeholder="Player 2 is X" />
-          <input id="startBtn" type="button" name="start" value="Start Game">
-        </div>
-        <div class="gameplay-container">
-          <div id="row-1" class="gameplay-row">
-            <div id="a1" class="gameplay-cell">
-            </div>
-            <div id="a2" class="gameplay-cell">
-            </div>
-            <div id="a3" class="gameplay-cell">
-            </div>
-          </div>
-          <div id="row-2" class="gameplay-row">
-            <div id="b1" class="gameplay-cell">
-            </div>
-            <div id="b2" class="gameplay-cell">
-            </div>
-            <div id="b3" class="gameplay-cell">
-            </div>
-          </div>
-          <div id="row-3" class="gameplay-row">
-            <div id="c1" class="gameplay-cell">
-            </div>
-            <div id="c2" class="gameplay-cell">
-            </div>
-            <div id="c3" class="gameplay-cell">
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <script src="jestIndex.js"></script>
-  </body>
-</html>
